Update and rename Actividades.jsx to ActividadadModal.jsx
